Move hazelcast implementation to a new module
