--- core/org.wso2.carbon.core/src/main/java/org/wso2/carbon/core/clustering/hazelcast/ParameterAdapter.java
@@ -1,63 +1,0 @@
-/*
-*  Copyright (c) 2005-2011, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-package org.wso2.carbon.core.clustering.hazelcast;
-
-import org.apache.axiom.om.OMElement;
-import org.apache.axis2.AxisFault;
-import org.apache.axis2.description.Parameter;
-import org.apache.axis2.description.ParameterInclude;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Adapter for {@link org.apache.axis2.description.ParameterInclude}
- */
-public class ParameterAdapter implements ParameterInclude {
-    protected final Map<String, Parameter> parameters = new HashMap<String, Parameter>();
-
-    public void addParameter(Parameter param) throws AxisFault {
-        parameters.put(param.getName(), param);
-    }
-
-    public void deserializeParameters(OMElement parameterElement) throws AxisFault {
-        throw new UnsupportedOperationException();
-    }
-
-    public Parameter getParameter(String name) {
-        return parameters.get(name);
-    }
-
-    public ArrayList getParameters() {
-        ArrayList<Parameter> list = new ArrayList<Parameter>();
-        for (String msg : parameters.keySet()) {
-            list.add(parameters.get(msg));
-        }
-        return list;
-    }
-
-    public boolean isParameterLocked(String parameterName) {
-        Parameter parameter = parameters.get(parameterName);
-        return parameter != null && parameter.isLocked();
-    }
-
-    public void removeParameter(Parameter param) throws AxisFault {
-        parameters.remove(param.getName());
-    }
-}
