--- core/org.wso2.carbon.core/src/main/java/org/wso2/carbon/core/clustering/hazelcast/aws/AWSConstants.java
@@ -1,35 +1,0 @@
-/*
-*  Copyright (c) 2005-2011, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-package org.wso2.carbon.core.clustering.hazelcast.aws;
-
-/**
- * TODO: class description
- */
-public class AWSConstants {
-
-    public static final String ACCESS_KEY = "accessKey";
-    public static final String SECRET_KEY = "secretKey";
-    public static final String SECURITY_GROUP = "securityGroup" ;
-    public static final String CONNECTION_TIMEOUT ="connTimeout" ;
-    public static final String HOST_HEADER = "hostHeader";
-    public static final String REGION = "region";
-    public static final String TAG_KEY = "tagKey";
-    public static final String TAG_VALUE = "tagValue";
-    public static final String IAM_ROLE = "iamRole";
-    public static final String NETWORK_INTERFACE = "networkInterface";
-}
