+++ core/org.wso2.carbon.hazelcast/src/main/java/org/wso2/carbon/hazelcast/wka/WKAConstants.java
@@ -1,0 +1,29 @@
+/*
+*  Copyright (c) 2005-2011, WSO2 LLC (http://www.wso2.org) All Rights Reserved.
+*
+*  WSO2 LLC licenses this file to you under the Apache License,
+*  Version 2.0 (the "License"); you may not use this file except
+*  in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied.  See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
+package org.wso2.carbon.hazelcast.wka;
+
+/**
+ * Constants used in implementation of Well-known Address membership scheme based on Hazelcast
+ */
+public final class WKAConstants {
+
+    public static final String CONNECTION_TIMEOUT = "connTimeout";
+
+    private WKAConstants() {
+    }
+}
